insert user
by admin

+++ Booking/database/migrations/2020_04_11_182520_create_addresses_table.php
@@ -20,9 +20,9 @@ class CreateAddressesTable extends Migration
             $table->string('customerLine2');
             $table->string('customerLine3');
             $table->string('postCode');
-            $table->integer('cityID')->unsigned();
+            $table->unsignedBigInteger('cityID')->unsigned();
             $table->foreign('cityID')->references('id')->on('cities');
-            $table->integer('countryID')->unsigned();
+            $table->unsignedBigInteger('countryID')->unsigned();
             $table->foreign('countryID')->references('id')->on('countries');
 
             $table->timestamps();
