+++ Booking/database/migrations/2020_04_11_182521_create_customers_table.php
@@ -19,9 +19,9 @@ class CreateCustomersTable extends Migration
             $table->string('firstName');
             $table->string('lastName');
             $table->date('dateOfBirth');
-            $table->integer('titleID')->unsigned();
+            $table->unsignedBigInteger('titleID')->unsigned();
             $table->foreign('titleID')->references('id')->on('titles');
-            $table->integer('customerAddressID')->unsigned();
+            $table->unsignedBigInteger('customerAddressID')->unsigned();
             $table->foreign('customerAddressID')->references('id')->on('addresses');
 
             $table->timestamps();
