+++ Booking/database/migrations/2020_04_11_183009_create_phones_table.php
@@ -15,11 +15,11 @@ class CreatePhonesTable extends Migration
     {
         Schema::create('phones', function (Blueprint $table) {
             $table->id();
-            $table->integer('phoneTypeID')->unsigned();
+            $table->unsignedBigInteger('phoneTypeID')->unsigned();
             $table->foreign('phoneTypeID')->references('id')->on('phone_types');
-            $table->integer('phoneExtentionID')->unsigned();
+            $table->unsignedBigInteger('phoneExtentionID')->unsigned();
             $table->foreign('phoneExtentionID')->references('id')->on('phone_extentions');
-            $table->integer('contactID')->unsigned();
+            $table->unsignedBigInteger('contactID')->unsigned();
             $table->foreign('contactID')->references('id')->on('contacts');
             $table->timestamps();
         });
