+++ Booking/database/migrations/2020_04_11_183749_create_properties_table.php
@@ -17,13 +17,13 @@ class CreatePropertiesTable extends Migration
             $table->id();
 
             $table->string('propertyName');
-            $table->integer('starRatingID')->unsigned();
+            $table->unsignedBigInteger('starRatingID')->unsigned();
             $table->foreign('starRatingID')->references('id')->on('official_star_ratings');
-            $table->integer('propertyHostID')->unsigned();
+            $table->unsignedBigInteger('propertyHostID')->unsigned();
             $table->foreign('propertyHostID')->references('id')->on('property_hosts');
-            $table->integer('propertyAddressID')->unsigned();
+            $table->unsignedBigInteger('propertyAddressID')->unsigned();
             $table->foreign('propertyAddressID')->references('id')->on('property_addresses');
-            $table->integer('propertyTypeID')->unsigned();
+            $table->unsignedBigInteger('propertyTypeID')->unsigned();
             $table->foreign('propertyTypeID')->references('id')->on('property_Types');
 
             $table->timestamps();
