+++ Booking/database/migrations/2020_04_11_183857_create_property_room_types_table.php
@@ -16,9 +16,9 @@ class CreatePropertyRoomTypesTable extends Migration
         Schema::create('property_room_types', function (Blueprint $table) {
             $table->id();
 
-            $table->integer('propertyID')->unsigned();
+            $table->unsignedBigInteger('propertyID')->unsigned();
             $table->foreign('propertyID')->references('id')->on('properties');
-            $table->integer('roomTypeID')->unsigned();
+            $table->unsignedBigInteger('roomTypeID')->unsigned();
             $table->foreign('roomTypeID')->references('id')->on('room_types');
 
             $table->timestamps();
