+++ Booking/database/migrations/2020_04_11_184040_create_property_facilities_table.php
@@ -16,9 +16,9 @@ class CreatePropertyFacilitiesTable extends Migration
         Schema::create('property_facilities', function (Blueprint $table) {
             $table->id();
 
-            $table->integer('propertyID')->unsigned();
+            $table->unsignedBigInteger('propertyID')->unsigned();
             $table->foreign('propertyID')->references('id')->on('properties');
-            $table->integer('facilityID')->unsigned();
+            $table->unsignedBigInteger('facilityID')->unsigned();
             $table->foreign('facilityID')->references('id')->on('facilities_types');
 
             $table->timestamps();
