+++ Booking/database/migrations/2020_04_11_184334_create_booking_infos_table.php
@@ -20,9 +20,9 @@ class CreateBookingInfosTable extends Migration
             $table->double('totalPrice');
             $table->integer('outstandingAmount');
             $table->integer('cancelled');
-            $table->integer('agentID')->unsigned();
+            $table->unsignedBigInteger('agentID')->unsigned();
             $table->foreign('agentID')->references('id')->on('agents');
-            $table->integer('customerID')->unsigned();
+            $table->unsignedBigInteger('customerID')->unsigned();
             $table->foreign('customerID')->references('id')->on('customers');
 
             $table->timestamps();
